id change to long

+++ auth/auth-model/src/main/java/com/meysam/auth/model/entity/Role.java
@@ -16,5 +16,5 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private BigDecimal Id;
+    private long Id;
 }
